+++ cmd2/argparse_custom.py
@@ -130,6 +130,7 @@ tokens with descriptions instead of just a table of tokens::
 
     The user sees this:
         ITEM_ID     Item Name
+        ============================
         1           My item
         2           Another item
         3           Yet another item
@@ -150,6 +151,7 @@ Since descriptive_header and CompletionItem.description are just strings, you
 can format them in such a way to have multiple columns::
 
     ITEM_ID     Item Name            Checked Out    Due Date
+    ==========================================================
     1           My item              True           02/02/2022
     2           Another item         False
     3           Yet another item     False