+++ tests_isolated/test_commandset/conftest.py
@@ -107,28 +107,6 @@ SHORTCUTS_TXT = """Shortcuts for other commands:
 @@: _relative_run_script
 """
 
-# Output from the show command with default settings
-SHOW_TXT = """allow_style: 'Terminal'
-debug: False
-echo: False
-editor: 'vim'
-feedback_to_output: False
-max_completion_items: 50
-quiet: False
-timing: False
-"""
-
-SHOW_LONG = """
-allow_style: 'Terminal'   # Allow ANSI text style sequences in output (valid values: Terminal, Always, Never)
-debug: False              # Show full traceback on exception
-echo: False               # Echo command issued into output
-editor: 'vim'             # Program used by 'edit'
-feedback_to_output: False # Include nonessentials in '|', '>' results
-max_completion_items: 50  # Maximum number of CompletionItems to display during tab completion
-quiet: False              # Don't print nonessential feedback
-timing: False             # Report execution times
-"""
-
 
 def normalize(block):
     """Normalize a block of text to perform comparison.